Change PlayTime field from uint to ulong

+++ src/CorePluginAPI/Core/Models/IPlayer.cs
@@ -7,7 +7,7 @@ public class PlayerData
     public string Name { get; set; } = "";
     public byte PlayerClass { get; set; }
     public byte SkillGroup { get; set; }
-    public uint PlayTime { get; set; }
+    public ulong PlayTime { get; set; }
     public byte Level { get; set; } = 1;
     public uint Experience { get; set; }
     public uint Gold { get; set; }
